Configuration overwrite keys are now applied to the default configuration when an ini file fails to be loaded. (#61)
Support usage without an ini configuration file on disk.  When an empty string is provided in the "ecal-ini-file" command line parameter default settings are used but the overwrite keys are ignored.  This change moves the processing of the overwrite keys to after the ini file load attempt has occurred; the overwire keys are now applied regardless of the success of loading an ini file.  This permits an application to specify an empty value for the "ecal-ini-file" and thus configure ecal entirely within the application.

Added the "/_vs_out" and ".vs/" folder to the .gitignore file to better support building using visual studio 2019's folder open cmake build flow (which works rather well presently).

Work around for daft Visual Studio 2019 (16.6.2) bug that causes build errors to be reported when there aren't any due to the presence of the word "error" in a cmake message.