Core/Udpcap: Fixed IP reassembly
Earlier versions of Udpcap had an issue with IP Reassembly. The issue actually was inside the used Pcap++ library, which would store fragments basically forever. The IPv4 ID however is only a 16 bit value, which overflows quite easily. So once this value overflows, the Pcap++ library will mix up old and new fragments and create faulty IP packets.

The issue was solved by wrapping the Pcap++ IPReassembly class with a custom one. The outer class keeps track of the last-use-timestamp for all packets that haven't been assembled, yet. When a packet is tool old it will be removed from the Pcap++ IPReassembly and not used any more.