Add clang-tidy support
The Bash script builds the repo and afterwards prepares for clang-tidy.
The Python script reads an array of excluded directories.
Each element itself is an array to be combined by the proper path separator.
It saves two JSON files, one for the included and one for the excluded directories.
The config file for clang-tidy is copied directly from Jason Turner's cpp_starter_project:
https://github.com/cpp-best-practices/cpp_starter_project.