5.11 Core: Refactored to prevent McAffee false positive (#1051)
This small refactoring solves a false-positive detection of McAffee in the following build configuration:

- VS 2015 Toolkit (v140)
- Win32
- Debug

Other build configuration were not affected. Please note that the detection was a false positive, there never was any harmful code in eCAL. But we tried to fix it here nevertheless, as McAffee otherwise prevents the usage of eCAL in that exact configuration.